requirements.txt
flask
flask-cors
google-genai
opencv-python-headless
numpy
pillow
supabase
python-dotenv
gunicorn
